Semana 6

+++ src/usuario/usuario.service.ts
@@ -1,7 +1,13 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  InternalServerErrorException,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { ListaUsuarioDTO } from './dto/ListaUsuario.dto';
 import { UsuarioEntity } from './usuario.entity';
+import { CriaUsuarioDTO } from './dto/CriaUsuario.dto';
 import { Repository } from 'typeorm';
 import { AtualizaUsuarioDTO } from './dto/AtualizaUsuario.dto';
 
@@ -13,7 +19,19 @@ export class UsuarioService {
   ) {}
 
   async criaUsuario(usuarioEntity: UsuarioEntity) {
-    await this.usuarioRepository.save(usuarioEntity);
+    const usuarioExistente = await this.buscaPorEmail(usuarioEntity.email);
+
+    if (usuarioExistente) {
+      throw new ConflictException('Este e-mail já está em uso.');
+    }
+
+    try {
+      await this.usuarioRepository.save(usuarioEntity);
+    } catch (error) {
+      throw new InternalServerErrorException(
+        'Erro ao criar usuário. Tente novamente mais tarde.',
+      );
+    }
   }
 
   async listUsuarios() {
@@ -28,14 +46,33 @@ export class UsuarioService {
     const checkEmail = await this.usuarioRepository.findOne({
       where: { email },
     });
+
+    if (checkEmail === null)
+      throw new NotFoundException('O email não foi encontrado.');
+
     return checkEmail;
   }
 
   async atualizaUsuario(id: string, novosDados: AtualizaUsuarioDTO) {
+    const usuarioExistente = await this.usuarioRepository.findOne({
+      where: { id },
+    });
+
+    if (!usuarioExistente) {
+      throw new BadRequestException('O usuário não foi encontrado.');
+    }
+
     await this.usuarioRepository.update(id, novosDados);
   }
 
   async deletaUsuario(id: string) {
+    const usuarioExistente = await this.usuarioRepository.findOne({
+      where: { id },
+    });
+
+    if (!usuarioExistente) {
+      throw new NotFoundException(`Usuário  não encontrado.`);
+    }
     await this.usuarioRepository.delete(id);
   }
 }
